+++ R/do_coords_minmax.R
@@ -1,4 +1,4 @@
-.do_coords_minmax <- function(df, reverse=""){
+.do_coords_minmax <- function(df, reverse.axis.values=""){
   res <- list("xmin" = floor(min(df$x,   na.rm = T) / 100) * 100,
               "xmax" = ceiling(max(df$x, na.rm = T) / 100) * 100,
               "ymin" = floor(min(df$y,   na.rm = T) / 100) * 100, 
@@ -7,7 +7,7 @@
               "zmax" = max(df$z, na.rm = T)
        )
   
-  if(grepl("z", reverse)){ names(res)[5:6] <- c("zmax", "zmin") }
+  if(grepl("z", reverse.axis.values)){ names(res)[5:6] <- c("zmax", "zmin") }
   
   res
 }